Fixes: 8616f552dec5 ("JavaWordToken is for parsing java reserved words")
The aforementioned commit tries to convert the text of a reserved word token to that of its corresponding enumerated type, but in doing so it first converts the token-text to lower case. All enumerators have all uppercase names by default, and as such the call to JavaTokenType.valueOf() fails saying it cannot find the enumerated type in question. For example: if JavaWordToken encounters the "public" reserved word token, it will convert it to lowercase (which it already is) and then call JavaTokenType.valueOf(), at which point JavaTokenType will look through its values for a value with the name "public" and fail to find it because the name of JavaTokenType.PUBLIC is not "public", it is "PUBLIC".